Remove TestCases namespace
