+++ Tests/Unit/Content/Domain/Model/TemplateTraitTest.php
@@ -16,12 +16,9 @@ namespace Sulu\Bundle\ContentBundle\Tests\Unit\Content\Domain\Model;
 use PHPUnit\Framework\TestCase;
 use Sulu\Bundle\ContentBundle\Content\Domain\Model\TemplateInterface;
 use Sulu\Bundle\ContentBundle\Content\Domain\Model\TemplateTrait;
-use Sulu\Bundle\ContentBundle\TestCases\Content\TemplateTestCaseTrait;
 
 class TemplateTraitTest extends TestCase
 {
-    use TemplateTestCaseTrait;
-
     protected function getTemplateInstance(): TemplateInterface
     {
         return new class() implements TemplateInterface {
@@ -32,5 +29,27 @@ class TemplateTraitTest extends TestCase
                 return 'example';
             }
         };
+    }
+
+    public function testGetSetTemplateKey(): void
+    {
+        $model = $this->getTemplateInstance();
+        $this->assertNull($model->getTemplateKey());
+        $model->setTemplateKey('template');
+        $this->assertSame('template', $model->getTemplateKey());
+    }
+
+    public function testGetSetTemplateData(): void
+    {
+        $model = $this->getTemplateInstance();
+        $this->assertSame([], $model->getTemplateData());
+        $model->setTemplateData(['data' => 'My Data']);
+        $this->assertSame(['data' => 'My Data'], $model->getTemplateData());
+    }
+
+    public function testGetTemplateType(): void
+    {
+        $model = $this->getTemplateInstance();
+        $this->assertSame('example', $model->getTemplateType());
     }
 }
